+++ apps/client/src/app/core/rxjs/of-packet-subtype.ts
@@ -1,10 +1,11 @@
 import { OperatorFunction } from 'rxjs';
 import { filter } from 'rxjs/operators';
+import { BasePacket } from '../../model/pcap/BasePacket';
 
-export function ofPacketSubType(
+export function ofPacketSubType<R extends BasePacket>(
   subType: string
-): OperatorFunction<any, any> {
-  return filter((packet: any) => {
+): OperatorFunction<BasePacket, R> {
+  return filter<R>((packet: BasePacket) => {
       return subType === packet.subType;
     }
   );