+++ apps/client/src/app/core/rxjs/of-packet-type.ts
@@ -1,11 +1,12 @@
 import { OperatorFunction } from 'rxjs';
 import { filter } from 'rxjs/operators';
+import { BasePacket } from '../../model/pcap/BasePacket';
 
-export function ofPacketType(
+export function ofPacketType<R extends BasePacket>(
   type: string,
   subType?: string
-): OperatorFunction<any, any> {
-  return filter((packet: any) => {
+): OperatorFunction<BasePacket, R> {
+  return filter<R>((packet: BasePacket) => {
       let matches = type === packet.type;
       if (packet.superType) {
         matches = matches || type === packet.superType;