+++ apps/client/src/app/model/pcap/EventPlay32.ts
@@ -1,0 +1,32 @@
+import { EventPlay4 } from './EventPlay4';
+
+export interface EventPlay32 extends EventPlay4 {
+  param5: number;
+  param6: number;
+  param7: number;
+  param8: number;
+  param9: number;
+  param10: number;
+  param11: number;
+  param12: number;
+  param13: number;
+  param14: number;
+  param15: number;
+  param16: number;
+  param17: number;
+  param18: number;
+  param19: number;
+  param20: number;
+  param21: number;
+  param22: number;
+  param23: number;
+  param24: number;
+  param25: number;
+  param26: number;
+  param27: number;
+  param28: number;
+  param29: number;
+  param30: number;
+  param31: number;
+  param32: number;
+}