+++ apps/client/src/app/pages/simulator/components/recipe-choice-popup/recipe-choice-popup.component.ts
@@ -7,7 +7,6 @@ import { GarlandToolsService } from '../../../../core/api/garland-tools.service'
 import { HtmlToolsService } from '../../../../core/tools/html-tools.service';
 import { debounceTime, filter, map, startWith, switchMap, tap } from 'rxjs/operators';
 import { TranslateService } from '@ngx-translate/core';
-import { Router } from '@angular/router';
 import { RotationPickerService } from '../../../../modules/rotations/rotation-picker.service';
 
 @Component({